# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      component:
        type: string
      created_at:
        description: 时间戳
        type: string
      external:
        type: boolean
      icon:
        type: string
      id:
        type: integer
      keep_alive:
        type: boolean
      meta:
        description: Meta 信息 (JSON)
        type: string
      name:
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/models.Menu'
        description: 关联
      parent_id:
        type: integer
      path:
        type: string
      permission:
        description: 权限相关
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      visible:
        type: integer
    type: object
  models.MenuMeta:
    properties:
      activeMenu:
        type: string
      affix:
        type: boolean
      alwaysShow:
        type: boolean
      breadcrumb:
        type: boolean
      hidden:
        type: boolean
      icon:
        type: string
      keepAlive:
        type: boolean
      noCache:
        type: boolean
      title:
        type: string
    type: object
  models.Role:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      name:
        type: string
      perm_list:
        items:
          type: string
        type: array
      status:
        type: integer
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_super_admin:
        description: Virtual field, not stored in database
        type: boolean
      nickname:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      status:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: Business status code
        type: integer
      data:
        description: Response data
      message:
        description: Response message
        type: string
      trace_id:
        description: Trace ID for request tracking
        type: string
    type: object
  services.CreateMenuRequest:
    properties:
      component:
        type: string
      external:
        type: boolean
      icon:
        type: string
      keep_alive:
        type: boolean
      meta:
        $ref: '#/definitions/models.MenuMeta'
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      permission:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      visible:
        type: integer
    required:
    - name
    - title
    type: object
...
  services.MenuMeta:
    properties:
      hidden:
        type: boolean
      icon:
        type: string
      keepAlive:
        type: boolean
      title:
        type: string
    type: object
  services.MenuRouteItem:
    properties:
      children:
        items:
          $ref: '#/definitions/services.MenuRouteItem'
        type: array
      component:
        type: string
      id:
        type: integer
      meta:
        $ref: '#/definitions/services.MenuMeta'
      name:
        type: string
      path:
        type: string
      sort:
        type: integer
    type: object
  services.UpdateMenuRequest:
    properties:
      component:
        type: string
      external:
        type: boolean
      icon:
        type: string
      keep_alive:
        type: boolean
      meta:
        $ref: '#/definitions/models.MenuMeta'
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      permission:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      visible:
        type: integer
    type: object
...
  v1.UpdateMenuRolesRequest:
    properties:
      role_ids:
        items:
          type: integer
        type: array
    required:
    - role_ids
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A modern Go admin scaffold API server.
  title: Go Admin Scaffold API
  version: "1.0"
paths:
...
  /admin/v1/menus:
    get:
      consumes:
      - application/json
      description: Get list of all menus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List menus
      tags:
      - menus
    post:
      consumes:
      - application/json
      description: Create a new menu
      parameters:
      - description: Menu data
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/services.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create menu
      tags:
      - menus
  /admin/v1/menus/{id}:
    delete:
      consumes:
      - application/json
      description: Delete menu by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete menu
      tags:
      - menus
    get:
      consumes:
      - application/json
      description: Get menu by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get menu
      tags:
      - menus
    put:
      consumes:
      - application/json
      description: Update menu by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu data
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/services.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update menu
      tags:
      - menus
  /admin/v1/menus/{id}/roles:
    put:
      consumes:
      - application/json
      description: Update roles associated with a menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role IDs
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateMenuRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update menu roles
      tags:
      - menus
  /admin/v1/menus/tree:
    get:
      consumes:
      - application/json
      description: Get menu tree structure
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get menu tree
      tags:
      - menus
  /admin/v1/menus/user:
    get:
      consumes:
      - application/json
      description: Get menus accessible by current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/services.MenuRouteItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get user menus
      tags:
      - menus
...
swagger: "2.0"